modified the project structure to manage different modules, also added the common module
