+++ common/pom.xml
@@ -2,16 +2,14 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>skt-task</artifactId>
+        <groupId>nearsoft.skytouch</groupId>
+        <version>1.0-SNAPSHOT</version>
+    </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>nearsoft.skytouch</groupId>
-    <artifactId>skt-task</artifactId>
-    <version>1.0-SNAPSHOT</version>
+    <artifactId>common</artifactId>
 
-    <packaging>pom</packaging>
-    <modules>
-        <module>management</module>
-        <module>common</module>
-    </modules>
 
 </project>
