+++ common/src/main/java/nearsoft/skytouch/common/model/Product.java
@@ -1,0 +1,32 @@
+package nearsoft.skytouch.common.model;
+
+public class Product {
+    private String name;
+    private String description;
+    private Long price;
+
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public Long getPrice() {
+        return price;
+    }
+
+    public void setPrice(Long price) {
+        this.price = price;
+    }
+}
