+++ management/src/main/java/nearsoft/skytouch/management/TaskApplication.java
@@ -1,0 +1,12 @@
+package nearsoft.skytouch.management;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TaskApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(TaskApplication.class, args);
+    }
+}
