+++ management/src/main/java/nearsoft/skytouch/management/controller/HomeController.java
@@ -1,0 +1,16 @@
+package nearsoft.skytouch.management.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+import java.util.Map;
+
+@Controller
+public class HomeController {
+    @RequestMapping("/")
+    public String welcome(Map<String, Object> model) {
+        return "/home";
+    }
+
+
+}
