+++ management/src/main/java/nearsoft/skytouch/management/repository/ProductRepository.java
@@ -1,0 +1,9 @@
+package nearsoft.skytouch.management.repository;
+
+import nearsoft.skytouch.common.model.Product;
+
+import java.util.List;
+
+public interface ProductRepository {
+    List<Product> retrieveProducts();
+}
