+++ management/src/main/java/nearsoft/skytouch/management/service/ProductService.java
@@ -1,0 +1,9 @@
+package nearsoft.skytouch.management.service;
+
+import nearsoft.skytouch.common.model.Product;
+
+import java.util.List;
+
+public interface ProductService {
+    List<Product> retrieveProducts();
+}
