+++ management/src/main/webapp/WEB-INF/jsp/home.jsp
@@ -1,0 +1,25 @@
+<%@ page language="java" contentType="text/html;charset=UTF-8"%>
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta http-equiv="Content-Type" content="text/html;charset=UTF-8">
+         <link href="webjars/bootstrap/3.3.6/css/bootstrap.min.css"
+                        rel="stylesheet">
+    </head>
+    <body>
+        <nav class="navbar navbar-inverse">
+          <div class="container-fluid">
+            <div class="navbar-header">
+              <a class="navbar-brand" href="/">Skytouch task</a>
+            </div>
+            <ul class="nav navbar-nav">
+              <li><a href="/products">Product list</a></li>
+              <li><a href="/view/product">Add product</a></li>
+            </ul>
+          </div>
+        </nav>
+        <h1>Skytouch task</h1>
+        <script src="webjars/jquery/1.9.1/jquery.min.js"></script>
+        <script src="webjars/bootstrap/3.3.6/js/bootstrap.min.js"></script>
+    </body>
+</html>
