+++ management/src/test/java/nearsoft/skytouch/management/TaskApplicationTests.java
@@ -1,0 +1,16 @@
+package nearsoft.skytouch.management;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class TaskApplicationTests {
+
+    @Test
+    public void contextLoads() {
+    }
+
+}
