--- src/main/java/nearsoft/skytouch/task/TaskApplication.java
@@ -1,12 +1,0 @@
-package nearsoft.skytouch.task;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TaskApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(TaskApplication.class, args);
-    }
-}
