--- src/main/java/nearsoft/skytouch/task/controller/HomeController.java
@@ -1,16 +1,0 @@
-package nearsoft.skytouch.task.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-import java.util.Map;
-
-@Controller
-public class HomeController {
-    @RequestMapping("/")
-    public String welcome(Map<String, Object> model) {
-        return "/home";
-    }
-
-
-}
