--- src/main/java/nearsoft/skytouch/task/repository/ProductRepository.java
@@ -1,9 +1,0 @@
-package nearsoft.skytouch.task.repository;
-
-import nearsoft.skytouch.task.model.Product;
-
-import java.util.List;
-
-public interface ProductRepository {
-    List<Product> retrieveProducts();
-}
