--- src/main/java/nearsoft/skytouch/task/service/ProductService.java
@@ -1,9 +1,0 @@
-package nearsoft.skytouch.task.service;
-
-import nearsoft.skytouch.task.model.Product;
-
-import java.util.List;
-
-public interface ProductService {
-    List<Product> retrieveProducts();
-}
