--- src/main/webapp/WEB-INF/jsp/home.jsp
@@ -1,10 +1,0 @@
-<%@ page language="java" contentType="text/html;charset=UTF-8"%>
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta http-equiv="Content-Type" content="text/html;charset=UTF-8">
-    </head>
-    <body>
-        <h1>Skytouch task</h1>
-    </body>
-</html>
