--- src/test/java/nearsoft/skytouch/task/TaskApplicationTests.java
@@ -1,16 +1,0 @@
-package nearsoft.skytouch.task;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class TaskApplicationTests {
-
-    @Test
-    public void contextLoads() {
-    }
-
-}
